requirements.txt
# Agent Practice - Dependencies

# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Task 3: AutoGen
pyautogen>=0.2.0

# Task 5: FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Optional: Advanced features
# langchain>=0.1.0
# huggingface_hub>=0.19.0
# openai>=1.0.0
